core: mv ProposalBytes into parameters mod
