+++ crates/node/src/tendermint_node.rs
@@ -6,7 +6,8 @@ use std::str::FromStr;
 use namada_apps_lib::cli::namada_version;
 use namada_apps_lib::config;
 pub use namada_apps_lib::tendermint_node::*;
-use namada_sdk::chain::{ChainId, ProposalBytes};
+use namada_sdk::chain::ChainId;
+use namada_sdk::parameters::ProposalBytes;
 use namada_sdk::storage::BlockHeight;
 use namada_sdk::time::DateTimeUtc;
 use thiserror::Error;
